Rework Mirage Delete
